Separated request serializers into different files
