+++ Tests/Shared/Core/JSONEncodingTests.swift
@@ -1,0 +1,251 @@
+//
+//  ParameterEncodingTests.swift
+//  VimeoNetworking
+//
+//  Created by Rogerio de Paula Assis on 9/2/19.
+//  Copyright © 2019 Vimeo. All rights reserved.
+//
+
+import Foundation
+import XCTest
+@testable import VimeoNetworking
+
+// MARK: -
+
+class JSONParameterEncodingTestCase: XCTestCase {
+    // MARK: Properties
+    
+    let urlRequest = URLRequest(url: URL(string: "https://example.com/")!)
+    let encoding = JSONEncoding.default
+    
+    // MARK: Tests
+    
+    func testJSONParameterEncodeNilParameters() {
+        do {
+            // Given, When
+            let URLRequest = try encoding.encode(self.urlRequest, with: nil)
+            
+            // Then
+            XCTAssertNil(URLRequest.url?.query, "query should be nil")
+            XCTAssertNil(URLRequest.value(forHTTPHeaderField: "Content-Type"))
+            XCTAssertNil(URLRequest.httpBody, "HTTPBody should be nil")
+        } catch {
+            XCTFail("Test encountered unexpected error: \(error)")
+        }
+    }
+    
+    func testJSONParameterEncodeComplexParameters() {
+        do {
+            // Given
+            let parameters: [String: Any] = [
+                "foo": "bar",
+                "baz": ["a", 1, true],
+                "qux": [
+                    "a": 1,
+                    "b": [2, 2],
+                    "c": [3, 3, 3]
+                ]
+            ]
+            
+            // When
+            let URLRequest = try encoding.encode(self.urlRequest, with: parameters)
+            
+            // Then
+            XCTAssertNil(URLRequest.url?.query)
+            XCTAssertNotNil(URLRequest.value(forHTTPHeaderField: "Content-Type"))
+            XCTAssertEqual(URLRequest.value(forHTTPHeaderField: "Content-Type"), "application/json")
+            XCTAssertNotNil(URLRequest.httpBody)
+            
+            if let httpBody = URLRequest.httpBody {
+                do {
+                    let json = try JSONSerialization.jsonObject(with: httpBody, options: .allowFragments)
+                    
+                    if let json = json as? NSObject {
+                        XCTAssertEqual(json, parameters as NSObject)
+                    } else {
+                        XCTFail("json should be an NSObject")
+                    }
+                } catch {
+                    XCTFail("json should not be nil")
+                }
+            } else {
+                XCTFail("json should not be nil")
+            }
+        } catch {
+            XCTFail("Test encountered unexpected error: \(error)")
+        }
+    }
+    
+    func testJSONParameterEncodeParametersRetainsCustomContentType() {
+        do {
+            // Given
+            var mutableURLRequest = URLRequest(url: URL(string: "https://example.com/")!)
+            mutableURLRequest.setValue("application/custom-json-type+json", forHTTPHeaderField: "Content-Type")
+            
+            let parameters = ["foo": "bar"]
+            
+            // When
+            let urlRequest = try encoding.encode(mutableURLRequest, with: parameters)
+            
+            // Then
+            XCTAssertNil(urlRequest.url?.query)
+            XCTAssertEqual(urlRequest.value(forHTTPHeaderField: "Content-Type"), "application/custom-json-type+json")
+        } catch {
+            XCTFail("Test encountered unexpected error: \(error)")
+        }
+    }
+}
+
+// TODO: AFNetworking tests to be ported [RDPA 02/09/2019]
+
+//- (void)testThatAFHTTPRequestSerializationSerializesPOSTRequestsProperly {
+//    NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//    request.HTTPMethod = @"POST";
+//
+//    NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:request withParameters:@{@"key":@"value"} error:nil];
+//    NSString *contentType = serializedRequest.allHTTPHeaderFields[@"Content-Type"];
+//
+//    XCTAssertNotNil(contentType);
+//    XCTAssertEqualObjects(contentType, @"application/x-www-form-urlencoded");
+//
+//    XCTAssertNotNil(serializedRequest.HTTPBody);
+//    XCTAssertEqualObjects(serializedRequest.HTTPBody, [@"key=value" dataUsingEncoding:NSUTF8StringEncoding]);
+//    }
+//
+//    - (void)testThatAFHTTPRequestSerializationSerializesPOSTRequestsProperlyWhenNoParameterIsProvided {
+//        NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//        request.HTTPMethod = @"POST";
+//
+//        NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:request withParameters:nil error:nil];
+//        NSString *contentType = serializedRequest.allHTTPHeaderFields[@"Content-Type"];
+//
+//        XCTAssertNotNil(contentType);
+//        XCTAssertEqualObjects(contentType, @"application/x-www-form-urlencoded");
+//
+//        XCTAssertNotNil(serializedRequest.HTTPBody);
+//        XCTAssertEqualObjects(serializedRequest.HTTPBody, [NSData data]);
+//        }
+//
+//        - (void)testThatAFHTTPRequestSerialiationSerializesQueryParametersCorrectly {
+//            NSURLRequest *originalRequest = [NSURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//            NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:originalRequest withParameters:@{@"key":@"value"} error:nil];
+//
+//            XCTAssertTrue([[[serializedRequest URL] query] isEqualToString:@"key=value"], @"Query parameters have not been serialized correctly (%@)", [[serializedRequest URL] query]);
+//            }
+//
+//            - (void)testThatEmptyDictionaryParametersAreProperlyEncoded {
+//                NSURLRequest *originalRequest = [NSURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//                NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:originalRequest withParameters:@{} error:nil];
+//                XCTAssertFalse([serializedRequest.URL.absoluteString hasSuffix:@"?"]);
+//                }
+//
+//                - (void)testThatAFHTTPRequestSerialiationSerializesURLEncodableQueryParametersCorrectly {
+//                    NSURLRequest *originalRequest = [NSURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//                    NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:originalRequest withParameters:@{@"key":@" :#[]@!$&'()*+,;=/?"} error:nil];
+//
+//                    XCTAssertTrue([[[serializedRequest URL] query] isEqualToString:@"key=%20%3A%23%5B%5D%40%21%24%26%27%28%29%2A%2B%2C%3B%3D/?"], @"Query parameters have not been serialized correctly (%@)", [[serializedRequest URL] query]);
+//                    }
+//
+//                    - (void)testThatAFHTTPRequestSerialiationSerializesURLEncodedQueryParametersCorrectly {
+//                        NSURLRequest *originalRequest = [NSURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//                        NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:originalRequest withParameters:@{@"key":@"%20%21%22%23%24%25%26%27%28%29%2A%2B%2C%2F"} error:nil];
+//
+//                        XCTAssertTrue([[[serializedRequest URL] query] isEqualToString:@"key=%2520%2521%2522%2523%2524%2525%2526%2527%2528%2529%252A%252B%252C%252F"], @"Query parameters have not been serialized correctly (%@)", [[serializedRequest URL] query]);
+//                        }
+//
+//                        - (void)testThatAFHTTPRequestSerialiationSerializesQueryParametersCorrectlyFromQuerySerializationBlock {
+//                            [self.requestSerializer setQueryStringSerializationWithBlock:^NSString *(NSURLRequest *request, NSDictionary *parameters, NSError *__autoreleasing *error) {
+//                                __block NSMutableString *query = [NSMutableString stringWithString:@""];
+//                                [parameters enumerateKeysAndObjectsUsingBlock:^(id key, id obj, BOOL *stop) {
+//                                [query appendFormat:@"%@**%@",key,obj];
+//                                }];
+//
+//                                return query;
+//                                }];
+//
+//                            NSURLRequest *originalRequest = [NSURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//                            NSURLRequest *serializedRequest = [self.requestSerializer requestBySerializingRequest:originalRequest withParameters:@{@"key":@"value"} error:nil];
+//
+//                            XCTAssertTrue([[[serializedRequest URL] query] isEqualToString:@"key**value"], @"Custom Query parameters have not been serialized correctly (%@) by the query string block.", [[serializedRequest URL] query]);
+//                            }
+//
+//                            - (void)testThatAFHTTPRequestSerialiationSerializesMIMETypeCorrectly {
+//                                NSMutableURLRequest *originalRequest = [NSMutableURLRequest requestWithURL:[NSURL URLWithString:@"http://example.com"]];
+//                                Class streamClass = NSClassFromString(@"AFStreamingMultipartFormData");
+//                                id <AFMultipartFormDataTest> formData = [[streamClass alloc] initWithURLRequest:originalRequest stringEncoding:NSUTF8StringEncoding];
+//
+//                                NSURL *fileURL = [NSURL fileURLWithPath:[[NSBundle bundleForClass:[self class]] pathForResource:@"ADNNetServerTrustChain/adn_0" ofType:@"cer"]];
+//
+//                                [formData appendPartWithFileURL:fileURL name:@"test" error:NULL];
+//
+//                                AFHTTPBodyPart *part = [formData.bodyStream.HTTPBodyParts firstObject];
+//
+//                                XCTAssertTrue([part.headers[@"Content-Type"] isEqualToString:@"application/x-x509-ca-cert"], @"MIME Type has not been obtained correctly (%@)", part.headers[@"Content-Type"]);
+//}
+//
+//#pragma mark -
+//
+//- (void)testThatValueForHTTPHeaderFieldReturnsSetValue {
+//    [self.requestSerializer setValue:@"Actual Value" forHTTPHeaderField:@"Set-Header"];
+//    NSString *value = [self.requestSerializer valueForHTTPHeaderField:@"Set-Header"];
+//    XCTAssertTrue([value isEqualToString:@"Actual Value"]);
+//    }
+//
+//    - (void)testThatValueForHTTPHeaderFieldReturnsNilForUnsetHeader {
+//        NSString *value = [self.requestSerializer valueForHTTPHeaderField:@"Unset-Header"];
+//        XCTAssertNil(value);
+//        }
+//
+//        - (void)testQueryStringSerializationCanFailWithError {
+//            AFHTTPRequestSerializer *serializer = [AFHTTPRequestSerializer serializer];
+//
+//            NSError *serializerError = [NSError errorWithDomain:@"TestDomain" code:0 userInfo:nil];
+//
+//            [serializer setQueryStringSerializationWithBlock:^NSString *(NSURLRequest *request, NSDictionary *parameters, NSError *__autoreleasing *error) {
+//                *error = serializerError;
+//                return nil;
+//                }];
+//
+//            NSError *error;
+//            NSURLRequest *request = [serializer requestWithMethod:@"GET" URLString:@"url" parameters:@{} error:&error];
+//            XCTAssertNil(request);
+//            XCTAssertEqual(error, serializerError);
+//            }
+//
+//            - (void)testThatHTTPHeaderValueCanBeRemoved {
+//                AFHTTPRequestSerializer *serializer = [AFHTTPRequestSerializer serializer];
+//                NSString *headerField = @"TestHeader";
+//                NSString *headerValue = @"test";
+//                [serializer setValue:headerValue forHTTPHeaderField:headerField];
+//                XCTAssertTrue([serializer.HTTPRequestHeaders[headerField] isEqualToString:headerValue]);
+//                [serializer setValue:nil forHTTPHeaderField:headerField];
+//                XCTAssertFalse([serializer.HTTPRequestHeaders.allKeys containsObject:headerField]);
+//}
+//
+//#pragma mark - Helper Methods
+//
+//- (void)testQueryStringFromParameters {
+//    XCTAssertTrue([AFQueryStringFromParameters(@{@"key":@"value",@"key1":@"value&"}) isEqualToString:@"key=value&key1=value%26"]);
+//    }
+//
+//    - (void)testPercentEscapingString {
+//        XCTAssertTrue([AFPercentEscapedStringFromString(@":#[]@!$&'()*+,;=?/") isEqualToString:@"%3A%23%5B%5D%40%21%24%26%27%28%29%2A%2B%2C%3B%3D?/"]);
+//}
+//
+//#pragma mark - #3028 tests
+////https://github.com/AFNetworking/AFNetworking/pull/3028
+//
+//- (void)testThatEmojiIsProperlyEncoded {
+//    //Start with an odd number of characters so we can cross the 50 character boundry
+//    NSMutableString *parameter = [NSMutableString stringWithString:@"!"];
+//    while (parameter.length < 50) {
+//        [parameter appendString:@"👴🏿👷🏻👮🏽"];
+//    }
+//
+//    AFHTTPRequestSerializer *serializer = [AFHTTPRequestSerializer serializer];
+//    NSURLRequest *request = [serializer requestWithMethod:@"GET"
+//        URLString:@"http://test.com"
+//        parameters:@{@"test":parameter}
+//        error:nil];
+//    XCTAssertTrue([request.URL.query isEqualToString:@"test=%21%F0%9F%91%B4%F0%9F%8F%BF%F0%9F%91%B7%F0%9F%8F%BB%F0%9F%91%AE%F0%9F%8F%BD%F0%9F%91%B4%F0%9F%8F%BF%F0%9F%91%B7%F0%9F%8F%BB%F0%9F%91%AE%F0%9F%8F%BD%F0%9F%91%B4%F0%9F%8F%BF%F0%9F%91%B7%F0%9F%8F%BB%F0%9F%91%AE%F0%9F%8F%BD%F0%9F%91%B4%F0%9F%8F%BF%F0%9F%91%B7%F0%9F%8F%BB%F0%9F%91%AE%F0%9F%8F%BD%F0%9F%91%B4%F0%9F%8F%BF%F0%9F%91%B7%F0%9F%8F%BB%F0%9F%91%AE%F0%9F%8F%BD"]);
+//}
